Added classes and changed how variables work
Working class declaration, class initialization,
class function definition, class function calling,
functions that have arguments but return nil are
now supported, floats can be converted to doubles,
fixed some bug in the tokenization when
dereferencing.

Normal variables are now pointers and need to be
dereferenced for normal access, this is because
the nth function and accessing class variables was
confusing. Dereference the variable when accessing
it's value normally for example when printing it
or when passing it as a function argument, but not
when calling the nth function, accessing class
variables or calling class functions. Function
variables are still normal and shouldn't be
dereferenced.